chore(backend): folder restructure in order to support multiple projects
